service/agents.py
from autogen_agentchat.agents import AssistantAgent, UserProxyAgent
import os
from dotenv import load_dotenv

# 加载 .env 文件中的环境变量
load_dotenv()

# 配置 LLM
# 确保你已经在 .env 文件中设置了 OPENAI_API_KEY
# 或者直接在这里提供你的 API 密钥和基础 URL（如果使用兼容 OpenAI 的 API）
# 例如:
# config_list = autogen.config_list_from_json(
#     "OAI_CONFIG_LIST",
#     filter_dict={
#         "model": ["gpt-4", "gpt-3.5-turbo"],
#     },
# )
# 或者手动配置:
api_key = os.getenv("OPENAI_API_KEY")
if not api_key:
    raise ValueError("OPENAI_API_KEY not found in .env file or environment variables.")

config_list = [
    {
        "model": "gpt-4", # 或者你选择的其他模型, e.g., gpt-3.5-turbo
        "api_key": api_key,
        # "base_url": "YOUR_API_BASE_URL" # 如果需要，取消注释并设置
    }
]

llm_config = {
    "config_list": config_list,
    "cache_seed": 42, # 用于缓存，设置为 None 禁用
    "temperature": 0.7,
}

# --- 通用智能体 ---

# 用户代理 - 代表用户发起请求和提供输入
user_proxy = UserProxyAgent(
    name="UserProxy",
    human_input_mode="NEVER", # 设置为 "TERMINATE" 或 "ALWAYS" 以允许人工输入
    max_consecutive_auto_reply=10,
    is_termination_msg=lambda x: x.get("content", "").rstrip().endswith("TERMINATE"),
    code_execution_config=False, # 设置为 {"work_dir": "coding"} 以允许执行代码
    system_message="""A human user or coordinator. You will initiate the conversation with a task request.
    Provide the necessary information, such as the essay topic or the essay to be revised.
    You might be asked for clarification or additional details.
    Signal the end of the task by replying with 'TERMINATE' in the end of your message.
    """
)

# --- 中文写作/修改相关智能体 ---

# 中文作文规划师
chinese_planner = AssistantAgent(
    name="ChinesePlanner",
    llm_config=llm_config,
    system_message="""你是一位中文作文写作规划师 (Chinese Essay Planner)。
    你的任务是根据用户提供的主题和要求，制定一个详细的中文作文写作大纲或修改计划。
    你需要考虑文章结构、论点、论据和语言风格。
    将最终的计划清晰地呈现出来，使用 Markdown 格式。不要写实际的作文内容，只负责规划。
    Plan the process step-by-step. If the user asks for revision, create a revision plan. If the user asks for writing, create a writing plan (outline).
    Reply 'TERMINATE' when your task is done.
    """
)

# 中文作文写手
chinese_writer = AssistantAgent(
    name="ChineseWriter",
    llm_config=llm_config,
    system_message="""你是一位经验丰富的中文作文写手 (Chinese Essay Writer)。
    你会根据规划师提供的大纲和要求，撰写一篇高质量的中文作文。
    注意语言流畅、逻辑清晰、表达准确。
    确保你的输出是完整的作文内容。
    Reply 'TERMINATE' when your task is done.
    """
)

# 中文作文评分/评估师 (可复用)
chinese_scorer = AssistantAgent(
    name="ChineseScorer",
    llm_config=llm_config,
    system_message="""你是一位严格的中文作文评分员 (Chinese Essay Scorer)。
    你的任务是评估给定的中文作文草稿或最终稿。
    你需要从内容相关性、结构逻辑、语言表达、思想深度等多个维度进行评分（例如，满分100分）。
    除了评分，你还需要提供具体的、有建设性的修改意见，指出优点和不足之处。
    请以清晰的 Markdown 格式输出评分和修改意见。
    Example Output:
    ```markdown
    **评分:** 85/100

    **优点:**
    *   论点清晰，紧扣主题。
    *   结构完整，段落过渡自然。

    **不足与修改建议:**
    *   部分语句表达略显啰嗦，建议精简。例如，将“我们必须认识到环境保护的重要性”改为“环境保护至关重要”。
    *   第三段论据稍显不足，建议补充具体实例或数据。
    *   结尾可以更有力，尝试升华主题。
    ```
    Reply 'TERMINATE' when your task is done.
    """
)

# 中文作文修改师
chinese_reviser = AssistantAgent(
    name="ChineseReviser",
    llm_config=llm_config,
    system_message="""你是一位专业的中文作文修改师 (Chinese Essay Reviser)。
    你会根据评分员的意见和原始作文，对中文作文进行修改和润色。
    你的目标是提升作文的整体质量，修正语法错误、改进表达、增强逻辑性。
    输出修改后的完整作文。
    Reply 'TERMINATE' when your task is done.
    """
)


# --- 英文写作/修改相关智能体 ---

# 英文作文规划师
english_planner = AssistantAgent(
    name="EnglishPlanner",
    llm_config=llm_config,
    system_message="""You are an English essay writing planner.
    Your task is to create a detailed outline or revision plan for an English essay based on the user's topic and requirements.
    Consider the essay structure (introduction, body paragraphs with topic sentences, conclusion), arguments, supporting evidence, and writing style (e.g., formal, informal).
    Present the final plan clearly using Markdown. Do not write the actual essay content, only plan.
    Plan the process step-by-step. If the user asks for revision, create a revision plan. If the user asks for writing, create a writing plan (outline).
    Reply 'TERMINATE' when your task is done.
    """
)

# 英文作文写手
english_writer = AssistantAgent(
    name="EnglishWriter",
    llm_config=llm_config,
    system_message="""You are an experienced English essay writer.
    You will write a high-quality English essay based on the outline and requirements provided by the planner.
    Pay attention to grammar, vocabulary, fluency, logical coherence, and accurate expression.
    Ensure your output is the complete essay content.
    Reply 'TERMINATE' when your task is done.
    """
)

# 英文作文评分/评估师 (可复用)
english_scorer = AssistantAgent(
    name="EnglishScorer",
    llm_config=llm_config,
    system_message="""You are a strict English essay scorer.
    Your task is to evaluate the given English essay draft or final version.
    You need to score it from multiple dimensions such as task achievement, coherence and cohesion, lexical resource, and grammatical range and accuracy (e.g., out of 100 points or using IELTS/TOEFL band descriptors).
    Besides scoring, you must provide specific, constructive feedback, pointing out strengths and weaknesses with examples.
    Output the score and feedback in a clear Markdown format.
    Example Output:
    ```markdown
    **Score:** 7.5/9.0 (IELTS Band)

    **Strengths:**
    *   Addresses all parts of the task.
    *   Good range of vocabulary used appropriately.
    *   Logical paragraphing and clear central topic in each paragraph.

    **Areas for Improvement & Suggestions:**
    *   Some grammatical errors related to articles ('a'/'an'/'the'). E.g., "Importance of **the** education..." should be "Importance of education...".
    *   Could use more varied sentence structures. Try incorporating complex sentences with subordinate clauses.
    *   Cohesion could be improved with more diverse linking words (e.g., 'Furthermore', 'Nevertheless' instead of repeating 'Also').
    ```
    Reply 'TERMINATE' when your task is done.
    """
)

# 英文作文修改师
english_reviser = AssistantAgent(
    name="EnglishReviser",
    llm_config=llm_config,
    system_message="""You are a professional English essay reviser.
    You will revise and polish the English essay based on the scorer's feedback and the original essay.
    Your goal is to improve the overall quality of the essay by correcting grammatical errors, enhancing vocabulary and sentence structure, and ensuring logical flow.
    Output the revised complete essay.
    Reply 'TERMINATE' when your task is done.
    """
)
