service/requirements.txt
python-dotenv
fastapi
uvicorn[standard] # standard includes websockets and other useful extras
sse-starlette # For Server-Sent Events streaming
